Create Default and Save
This project aims to promote personal discipline and regulation of internet use. 
This is to reduce the negative impacts of this technology and internet driven age that we we live in.

It has two main core:
Create a discipline and social reward of better relationships with ourselves offline. 
And create more savings and investment when a user defaults. 

Thank you for any contribution to this project.